完成部分json通信

+++ back_end/v1.3 json/Game_Logic/json_io.py
@@ -22,6 +22,7 @@ def json_writer(file_name, data):
     :param data:
     :return:
     """
+    if not os.path.isfile(file_name):
+        raise IOError("JSON file does not exist.")
     with open(file_name, 'w') as f:
         json.dump(data, f, indent=4)
-
